request data unpacked in server process
