+++ tests/test_logic.py
@@ -80,18 +80,3 @@ class TestLogic(unittest.TestCase):
             for value in arrangement:
                 Utilities.change_board_value(test_board.board_data, value, test_marker)
             self.assertTrue(Logic.win_check(test_board, test_marker))
-
-    def test_server_process(self):
-        test_board = Board()
-        test_board.create_board(3)
-        test_player = HumanPlayer(["Marx", "human", "X"])
-        test_move = "1"
-        request_data = Utilities.generate_payload(test_board, test_player, test_move)
-        response_data = Logic.server_process(request_data)
-        self.assertTrue(response_data["move_success"])
-        self.assertEqual(
-            Utilities.check_board_value(
-                response_data["game_data"]["board"]["board_state"], "1"
-            ),
-            "X",
-        )
